// src/types/AmenityType.ts
export type AmenityType = 
  | "Free Wi-Fi"
  | "Fitness Center"
  | "Swimming Pool"
  | "Free Breakfast"
  | "TV"
  | "Business Center Access"
  | "Meeting Room"
  | "Complimentary Breakfast"
  | "Adjoining Rooms"
  | "Kitchenette"
  | "Play Area"
  | "Budget-Friendly"
  | "Single Bed"
  | "King Size Bed"
  | "City View"
  | "Room Service"
  | "Jacuzzi"
  | "Mini Bar"
  | "Ocean View"
  | "wifi"                
  | "Private Balcony"     
  | "Fireplace"          
  | "Hiking Trails"       
  | "Spa Services"
  |"Ocean View Balcony";       